+++ shaders/makeGoals.metal
@@ -5,7 +5,7 @@
 using namespace metal;
 
 kernel void makeGoals(
-	const uint2 tid [[thread_position_in_grid]],
+	const ushort2 tid [[thread_position_in_grid]],
 	const constant Uniforms& uniforms [[buffer(0)]],
 	const constant Pixel* previousTick [[buffer(1)]],
 	device Goal* goals [[buffer(2)]]
@@ -20,5 +20,5 @@ kernel void makeGoals(
 	
 	const Goal myGoal = previous.goalForCellAt(position, uniforms.frameNumber);
 	
-	goals[uint(position.y) * uint(uniforms.width) + uint(position.x)] = myGoal;
+	goals[int(position.y) * int(uniforms.width) + int(position.x)] = myGoal;
 }
