+++ shaders/updatePixels.metal
@@ -6,7 +6,7 @@
 using namespace metal;
 
 kernel void updatePixels(
-	const uint2 tid [[thread_position_in_grid]],
+	const ushort2 tid [[thread_position_in_grid]],
 	const constant Uniforms& uniforms [[buffer(0)]],
 	const constant Pixel* previousTick [[buffer(1)]],
 	device Pixel* currentTick [[buffer(2)]],
@@ -21,7 +21,7 @@ kernel void updatePixels(
 	const InputBoard previous { previousTick, uniforms };
 	OutputBoard current { currentTick, uniforms };
 	
-	const Goal myGoal = goals[uint(position.y) * uint(uniforms.width) + uint(position.x)];
+	const Goal myGoal = goals[int(position.y) * int(uniforms.width) + int(position.x)];
 	
 	switch (myGoal.kind) {
 		case Goal::Kind::change: {
